Removed old commented-out code
oneone.c, tiles.c, and tiles.h all had code from a while ago that has been commented out for a while that I've removed to clean stuff up. I might have to refer to the previous version of oneone.c if I ever need to see how the level was copied, but for now it isn't needed.