Commented out second Mario update and compensated
Finally fixed the issue where everything get's slow if you jump. Still not sure why it's just when you're jumping but it works. Also, at some point I need to CORRECTLY compensate all of the speeds because at the moment they're all like "dx += 0x1f * 2" which is a horrible way to do it but whatever it works.