Finally finished horizontal response
FINALLY holy cow this took forever. Most of the time working on this has been spent blankly staring at my code trying to figure out what was wrong. Eventually a combination of changing random stuff and hoping nothing breaks resulted in fixing this bug. I'll probably have to polish this a bit better and clean up all of my experimental code but it works for now.